refractoring: new api to delete goal with its childrens
